requirements.txt
discord.py==1.7.3
pymongo==4.2.0
requests==2.28.1
lomond==0.3.3
dnspython==2.0.0
pytz==2022.4
aniso8601==9.0.1
unidecode==1.3.6
bs4==0.0.1
jwt==1.3.1
websockets==10.3